pyproject.toml
[tool.poetry]
name = "toptracker-api"
version = "0.1.0"
description = ""
authors = ["Milan Slesarik <milslesarik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
